requirements.txt
mysql-connector-python
sqlalchemy
fastapi
uvicorn
python-dotenv
yt-dlp
ffmpeg-python
pytesseract
openai-whisper
pydub
pyaudio
boto3
numpy
pytesseract
Pillow
ImageHash
googlemaps
openai
python-decouple